+++ routes/auth.py
@@ -1,0 +1,76 @@
+from flask import Blueprint, render_template, request, redirect, url_for, session
+from werkzeug.security import generate_password_hash, check_password_hash
+from functools import wraps
+from database.db import get_db_connection
+
+auth_bp = Blueprint('auth', __name__)
+
+def login_required(f):
+    @wraps(f)
+    def decorated_function(*args, **kwargs):
+        if 'user_id' not in session:
+            return redirect(url_for('auth.login'))
+        return f(*args, **kwargs)
+    return decorated_function
+
+@auth_bp.route('/login', methods=['GET', 'POST'])
+def login():
+    if 'user_id' in session:
+        next_page = request.args.get('next')
+        if next_page and next_page.startswith('/'):
+            return redirect(next_page)
+        return redirect(url_for('rankings.rankings'))
+    
+    if request.method == 'POST':
+        username = request.form['username']
+        password = request.form['password']
+        
+        conn = get_db_connection()
+        user = conn.execute('SELECT * FROM users WHERE username = ?', 
+                          (username,)).fetchone()
+        conn.close()
+        
+        if user and check_password_hash(user['password'], password):
+            session['user_id'] = user['id']
+            session['username'] = user['username']
+            
+            next_page = request.args.get('next')
+            if next_page and next_page.startswith('/'):
+                return redirect(next_page)
+            return redirect(url_for('rankings.rankings'))
+        
+        return render_template('login.html', error='Invalid username or password')
+    
+    return render_template('login.html')
+
+@auth_bp.route('/signup', methods=['GET', 'POST'])
+def signup():
+    if request.method == 'POST':
+        username = request.form['username']
+        password = request.form['password']
+        confirm_password = request.form['confirm_password']
+        
+        if password != confirm_password:
+            return render_template('signup.html', error='Passwords do not match')
+        
+        conn = get_db_connection()
+        
+        if conn.execute('SELECT id FROM users WHERE username = ?', 
+                       (username,)).fetchone():
+            conn.close()
+            return render_template('signup.html', error='Username already exists')
+        
+        hashed_password = generate_password_hash(password)
+        conn.execute('INSERT INTO users (username, password) VALUES (?, ?)',
+                    (username, hashed_password))
+        conn.commit()
+        conn.close()
+        
+        return redirect(url_for('auth.login'))
+    
+    return render_template('signup.html')
+
+@auth_bp.route('/logout')
+def logout():
+    session.clear()
+    return redirect(url_for('auth.login'))
