+++ templates/base.html
@@ -21,17 +21,17 @@
             <div class="collapse navbar-collapse" id="navbarContent">
                 <ul class="navbar-nav me-auto">
                     <li class="nav-item">
-                        <a class="nav-link text-white-50 {% if request.endpoint == 'rankings' %}active text-white{% endif %}" 
-                           href="{{ url_for('rankings') }}">Rankings</a>
+                        <a class="nav-link text-white-50 {% if request.endpoint == 'rankings.rankings' %}active text-white{% endif %}" 
+                           href="{{ url_for('rankings.rankings') }}">Rankings</a>
                     </li>
                     {% if session.get('user_id') %}
                     <li class="nav-item">
-                        <a class="nav-link text-white-50 {% if request.endpoint == 'draw' %}active text-white{% endif %}" 
-                           href="{{ url_for('draw') }}">Draw</a>
+                        <a class="nav-link text-white-50 {% if request.endpoint == 'draw.draw' %}active text-white{% endif %}" 
+                           href="{{ url_for('draw.draw') }}">Draw</a>
                     </li>
                     <li class="nav-item">
-                        <a class="nav-link text-white-50 {% if request.endpoint == 'history' %}active text-white{% endif %}" 
-                           href="{{ url_for('history') }}">History</a>
+                        <a class="nav-link text-white-50 {% if request.endpoint == 'history.history' %}active text-white{% endif %}" 
+                           href="{{ url_for('history.history') }}">History</a>
                     </li>
                     {% endif %}
                 </ul>
@@ -42,11 +42,11 @@
                             <span class="nav-link text-white-50 small">{{ session.get('username') }}</span>
                         </li>
                         <li class="nav-item order-lg-2 order-1">
-                            <a class="nav-link text-danger small" href="{{ url_for('logout') }}">Logout</a>
+                            <a class="nav-link text-danger small" href="{{ url_for('auth.logout') }}">Logout</a>
                         </li>
                     {% else %}
                         <li class="nav-item">
-                            <a class="nav-link text-white-50 small" href="{{ url_for('login') }}">Login</a>
+                            <a class="nav-link text-white-50 small" href="{{ url_for('auth.login') }}">Login</a>
                         </li>
                     {% endif %}
                 </ul>
