+++ templates/login.html
@@ -22,7 +22,7 @@
                     {% if error %}
                     <div class="alert alert-danger">{{ error }}</div>
                     {% endif %}
-                    <form method="POST" action="{{ url_for('login') }}" class="auth-form">
+                    <form method="POST" action="{{ url_for('auth.login') }}" class="auth-form">
                         <div class="mb-3">
                             <label for="username" class="form-label">Username</label>
                             <input type="text" class="form-control" id="username" name="username" required>
@@ -34,7 +34,7 @@
                         <button type="submit" class="btn btn-primary w-100">Login</button>
                     </form>
                     <p class="text-center mt-4">
-                        Don't have an account? <a href="{{ url_for('signup') }}" class="auth-link">Sign up</a>
+                        Don't have an account? <a href="{{ url_for('auth.signup') }}" class="auth-link">Sign up</a>
                     </p>
                 </div>
             </div>
